:recycle: [FolderTreeDownloader] Refacto folder download using specific domain
