+++ src/ManagerV2/FolderManager.php
@@ -5,21 +5,13 @@ namespace App\ManagerV2;
 use App\Api\Manager\ApiClientManager;
 use App\Entity\Beneficiaire;
 use App\Entity\Client;
-use App\Entity\Document;
 use App\Entity\Dossier;
 use App\Repository\DossierRepository;
-use App\ServiceV2\BucketService;
 use App\ServiceV2\Traits\UserAwareTrait;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\EntityManagerInterface;
-use Psr\Log\LoggerInterface;
 use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Component\HttpFoundation\HeaderUtils;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use ZipStream\Exception\OverflowException;
-use ZipStream\ZipStream;
 
 class FolderManager
 {
@@ -27,8 +19,6 @@ class FolderManager
 
     public function __construct(
         private readonly EntityManagerInterface $em,
-        private readonly BucketService $bucketService,
-        private readonly LoggerInterface $logger,
         private readonly TranslatorInterface $translator,
         private readonly DossierRepository $folderRepository,
         private readonly Security $security,
@@ -69,52 +59,6 @@ class FolderManager
         $this->em->remove($folder);
 
         $this->em->flush();
-    }
-
-    public function getZipFromFolder(Dossier $folder): ?StreamedResponse
-    {
-        return 0 === $folder->getDocuments()->count()
-            ? null
-            : new StreamedResponse(
-                fn () => $this->createZipFromFolder($folder),
-                Response::HTTP_OK,
-                [
-                    'Content-Type' => 'application/zip',
-                    'Content-Disposition' => HeaderUtils::makeDisposition(
-                        HeaderUtils::DISPOSITION_ATTACHMENT,
-                        sprintf('%s.zip', $folder->getNom()),
-                    ),
-                ]
-            );
-    }
-
-    private function createZipFromFolder(Dossier $folder): void
-    {
-        $zip = new ZipStream(outputName: sprintf('%s.zip', $folder->getNom()));
-
-        $documents = $folder->getDocuments($this->getUser() === $folder->getBeneficiaire()->getUser())
-            ->filter(fn (Document $document) => is_resource($this->bucketService->getObjectStream($document->getObjectKey())));
-
-        foreach ($documents as $document) {
-            $objectStream = $this->bucketService->getObjectStream($document->getObjectKey());
-            if ($objectStream) {
-                $zip->addFileFromStream(
-                    $document->getNom(),
-                    $objectStream,
-                );
-            }
-        }
-
-        try {
-            $zip->finish();
-        } catch (OverflowException $e) {
-            $this->logger->error(sprintf(
-                'Error during zip download for folder %d from beneficiary %d, cause %s',
-                $folder->getId(),
-                $folder->getBeneficiaire()->getId(),
-                $e->getMessage(),
-            ));
-        }
     }
 
     /**
